Split out LibGit2Sharp functionality from ILocalGitClient
